Create utility method for inventory controller
